move functions to CollectionsDialog
