+++ src/components/Icons/CollectionRowActionIcon.js
@@ -5,11 +5,10 @@ import {
 import { Entypo } from '@expo/vector-icons'; // eslint-disable-line import/no-extraneous-dependencies
 import { connect } from 'react-redux';
 import {
-  arrayOf, func, number, string,
+  number, string,
 } from 'prop-types';
 
-import { deleteCollection, renameCollection, duplicateCollection } from '../../redux/action-creators';
-import { collectionsCountSelector, collectionsLabelsSelector } from '../../redux/selectors';
+import { collectionsCountSelector } from '../../redux/selectors';
 import Colors from '../../constants/Colors';
 import CollectionsDialog, { CollectionDialogText } from '../Dialog/CollectionsDialog';
 
@@ -17,21 +16,8 @@ const CollectionRowActionIcon = ({
   collectionId,
   collectionLabel,
   collectionsCount,
-  deleteCollectionAction,
-  renameCollectionAction,
-  duplicateCollectionAction,
-  collectionsLabels,
 }) => {
   const [collectionDialogText, setCollectionDialogText] = useState(null);
-  const [showUniqueError, setShowUniqueError] = useState(false);
-
-  const checkUniqueName = ({ inputText, rename, collectionLabel: label }) => {
-    // if action is rename, new label can be same as old label
-    if (rename && (inputText === label)) {
-      return true;
-    }
-    return !collectionsLabels.includes(inputText);
-  };
 
   const handlePress = () => {
     ActionSheetIOS.showActionSheetWithOptions(
@@ -59,24 +45,6 @@ const CollectionRowActionIcon = ({
     );
   };
 
-  const handleSubmit = (inputText) => {
-    if (checkUniqueName({ inputText, rename: collectionDialogText.action === 'rename', collectionLabel })) {
-      if (collectionDialogText.action === 'rename') {
-        renameCollectionAction(collectionId, inputText);
-      } if (collectionDialogText.action === 'duplicate') {
-        duplicateCollectionAction(collectionId, inputText);
-      } if (collectionDialogText.action === 'delete') {
-        deleteCollectionAction(collectionId);
-      } if (collectionDialogText.action === 'deleteError') {
-        setCollectionDialogText(null);
-      }
-      setCollectionDialogText(null);
-      setShowUniqueError(false);
-    } else {
-      setShowUniqueError(true);
-    }
-  };
-
   return (
     <View>
       <TouchableOpacity onPress={handlePress}>
@@ -84,11 +52,10 @@ const CollectionRowActionIcon = ({
       </TouchableOpacity>
       {collectionDialogText && (
       <CollectionsDialog
+        collectionId={collectionId}
+        collectionLabel={collectionLabel}
         collectionDialogText={collectionDialogText}
         setCollectionDialogText={setCollectionDialogText}
-        showUniqueError={showUniqueError}
-        handleSubmit={handleSubmit}
-        collectionLabel={collectionLabel}
       />
       )}
     </View>
@@ -99,21 +66,10 @@ CollectionRowActionIcon.propTypes = {
   collectionId: string.isRequired,
   collectionLabel: string.isRequired,
   collectionsCount: number.isRequired,
-  deleteCollectionAction: func.isRequired,
-  renameCollectionAction: func.isRequired,
-  duplicateCollectionAction: func.isRequired,
-  collectionsLabels: arrayOf(string.isRequired).isRequired,
 };
 
 const mapStateToProps = (state) => ({
   collectionsCount: collectionsCountSelector(state),
-  collectionsLabels: collectionsLabelsSelector(state),
 });
 
-const mapDispatchToProps = {
-  deleteCollectionAction: deleteCollection,
-  renameCollectionAction: renameCollection,
-  duplicateCollectionAction: duplicateCollection,
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(CollectionRowActionIcon);
+export default connect(mapStateToProps, null)(CollectionRowActionIcon);
